# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tournament Management API",
        "description": "Tournament Management API for Android App. This API provides endpoints for managing tournaments, user authentication, and player interactions.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/organizer/dashboard": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Get organizer dashboard data",
                "description": "REQUIRES: Authorization header with Bearer token. Format: 'Bearer {token}'. Example: 'Bearer 6|YuxRBo5pyMcpMOraGyRKfNZnqKsVuwQ7dG5jKu4Dab7b101f'",
                "operationId": "f17a17528f31a7f12e5f0a1914b4331a",
                "responses": {
                    "200": {
                        "description": "Dashboard data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tournaments": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "stats": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Missing or invalid Bearer token"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/tournaments": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "List all tournaments for organizer",
                "operationId": "d9ec984ceb3d2c8b08d40e49694184de",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "draft",
                                "open",
                                "active",
                                "expired",
                                "published"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tournaments"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Create a new tournament",
                "operationId": "98adc0975c6f70a14626cfafa22cec41",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sports_category_id",
                                    "team_name",
                                    "location",
                                    "start_date",
                                    "winning_date",
                                    "slot_count",
                                    "rules",
                                    "entry_fee"
                                ],
                                "properties": {
                                    "sports_category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "team_name": {
                                        "type": "string",
                                        "example": "Team Alpha"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Stadium A"
                                    },
                                    "location_details": {
                                        "type": "string",
                                        "example": "Near City Center"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-02-01"
                                    },
                                    "winning_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-02-15"
                                    },
                                    "slot_count": {
                                        "type": "integer",
                                        "example": 16
                                    },
                                    "template": {
                                        "type": "string",
                                        "example": "Template details"
                                    },
                                    "rules": {
                                        "type": "string",
                                        "example": "Tournament rules"
                                    },
                                    "entry_fee": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "price_details": {
                                        "type": "string",
                                        "example": "Price breakdown"
                                    },
                                    "ball_type": {
                                        "type": "string",
                                        "example": "Football"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tournament created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/tournaments/{id}": {
            "get": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Get tournament details",
                "operationId": "7405a95a8152cc70366905835e26f370",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Update tournament",
                "operationId": "c41c08f0b942ef55b9427fc5d4e6d4cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tournament updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/organizer/tournaments/{id}/publish": {
            "post": {
                "tags": [
                    "Organizer"
                ],
                "summary": "Publish tournament",
                "operationId": "9931a0954dfb6505a2f4f920a3a2d50a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament published successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "REQUIRED: Bearer token authentication. Format: \"Bearer {token}\". Example: \"Bearer 6|YuxRBo5pyMcpMOraGyRKfNZnqKsVuwQ7dG5jKu4Dab7b101f\". You MUST include the word \"Bearer\" followed by a space before your token.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}